+++ tests/files-test.js
@@ -3,6 +3,7 @@ import {
     writeTextInFile,
     concatenateFiles,
     copyFile,
+    copyDirectory,
     deleteFile
 } from "../files.js";
 
@@ -13,13 +14,16 @@ const a = `a`;
 const b = `b`;
 const pathDeep = `${a}/${b}/c.txt`;
 
-writeTextInFile(pathDeep, initialString).then(function () {
-    textFileContent(pathDeep).then(async function (result) {
-        console.log(`success`, result === initialString);
-        console.log(result);
+// writeTextInFile(pathDeep, initialString).then(function () {
+//     textFileContent(pathDeep).then(async function (result) {
+//         console.log(`success`, result === initialString);
+//         console.log(result);
 
-        await deleteFile(pathDeep);
-    });
-});
+//         await deleteFile(pathDeep);
+//     });
+// });
+
+copyDirectory(`tests`, `myCopy`)
 
 
+// copyFile(`tests/files-test.js`, `copyFile/files-test.js`)